fixed missed float.  and changed display to %g

+++ Rectangles/Rectangles/XYPoint.h
@@ -10,7 +10,7 @@
 
 @interface XYPoint: NSObject 
 
-@property int x, y;
+@property float x, y;
 
 -(void) setX: (float) xVal andY: (float) yVal; 
 @end
